+++ sinks/interamap/config.py
@@ -1,1 +1,1 @@
-ONLINE_MODE = True
+ONLINE_MODE = False
